Upgrade nuget, fix funcs

+++ src/BabyKusto.Core/Evaluation/BuiltIns/ScalarFunctions/GetMonthFunction.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 
-
 namespace BabyKusto.Core.Evaluation.BuiltIns.Impl;
 
 [KustoImplementation]
 internal class GetMonthFunction
 {
-    private static long Impl(DateTime date) => date.Month;
+    private static int Impl(DateTime date) => date.Month;
 }
