+++ src/BabyKusto.Core/Evaluation/BuiltIns/ScalarFunctions/GetYearFunction.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 
-
 namespace BabyKusto.Core.Evaluation.BuiltIns.Impl;
 
 [KustoImplementation]
 internal class GetYearFunction
 {
-    private static long Impl(DateTime date) => date.Year;
+    private static int Impl(DateTime date) => date.Year;
 }
